libutils: Add memcpy-family and printf functions
Copy over memzero and memset from the seL4 kernel. memcpy has to be CHERI-aware,
so it is copied from CHERI-newlib.

printf is the same from the Microkit's VMM example, it is copied as is but
also add support for printing formatted CHERI capabilities, this code is
copied from Arm's Morello muslc and is modified to work for standard
CHERI-RISC-V.

+++ build_sdk.py
@@ -669,12 +669,13 @@ def build_lib_component(
     # Make output read-only
     dest.chmod(0o744)
 
-    link_script = Path(component_name) / "microkit.ld"
-    dest = lib_dir / "microkit.ld"
-    dest.unlink(missing_ok=True)
-    copy(link_script, dest)
-    # Make output read-only
-    dest.chmod(0o744)
+    if component_name == "libmicrokit":
+        link_script = Path(component_name) / "microkit.ld"
+        dest = lib_dir / "microkit.ld"
+        dest.unlink(missing_ok=True)
+        copy(link_script, dest)
+        # Make output read-only
+        dest.chmod(0o744)
 
     include_dir = root_dir / "board" / board.name / config.name / "include"
     source_dir = Path(component_name) / "include"
@@ -834,8 +835,10 @@ def main() -> None:
             # A purecap protection domain will use libmicrokit_purecap.a while others will use libmicrokit.a
             if config.cheri:
                 build_lib_component("libmicrokit", root_dir, build_dir, board, config, args.llvm, True)
+                build_lib_component("libutils", root_dir, build_dir, board, config, args.llvm, True)
 
             build_lib_component("libmicrokit", root_dir, build_dir, board, config, args.llvm, False)
+            build_lib_component("libutils", root_dir, build_dir, board, config, args.llvm, False)
 
     # Setup the examples
     for example, example_path in EXAMPLES.items():
